I'm trying a new strategy for makeAllSums
I take the total sum, n(n+1)/2, and I either subtract the current iteration's value of n or I keep it the same, while adding each value of partialSum to the arrayList. The problem is that in fibonacci and the containSum, there was an operation that linked the two branches of the tree when it split off: the plus and or operator. If I'm just adding values to the arrayList then I can't join the two recursive calls.