I'm going to use a groupSum as a helper
makeAllSums makes an array list of 1-n. Then it goes though a for loop ending when i = n^2. For every i, it will call groupSum(0, nums, i) to check if i is a possible sum. If it is we add to the output arrayList and exiting the for loop we return it.